refactor ItemHeaderActions
