+++ .changeset/dry-buttons-build.md
@@ -1,0 +1,5 @@
+---
+'@keystonejs/app-admin-ui': patch
+---
+
+Refactored itemHeaderAction to export components instead of passing as parameter
